change payment amount type and generate with go-jet

+++ db/postgres/migrations/00002_create_payments_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS payments (
     id VARCHAR(36) PRIMARY KEY NOT NULL UNIQUE,
     payment_method_id VARCHAR(2) NOT NULL REFERENCES payment_methods(id),
-    amount INT NOT NULL,
+    amount BIGINT NOT NULL,
     description VARCHAR(255) NOT NULL,
     status VARCHAR(255) NOT NULL,
     callback_url VARCHAR(1024) NOT NULL,
